+++ XmlFormatter.php
@@ -40,6 +40,17 @@ class XmlFormatter extends Object implements FormatterInterface
      * @var string the name of the root element.
      */
     public $rootTag = 'request';
+    /**
+     * @var string the name of the elements that represent the array elements with numeric keys.
+     * @since 2.0.1
+     */
+    public $itemTag = 'item';
+    /**
+     * @var boolean whether to interpret objects implementing the [[\Traversable]] interface as arrays.
+     * Defaults to `true`.
+     * @since 2.0.1
+     */
+    public $useTraversableAsArray = true;
 
 
     /**
@@ -79,19 +90,9 @@ class XmlFormatter extends Object implements FormatterInterface
      */
     protected function buildXml($element, $data)
     {
-        if (is_object($data)) {
-            $child = new DOMElement(StringHelper::basename(get_class($data)));
-            $element->appendChild($child);
-            if ($data instanceof Arrayable) {
-                $this->buildXml($child, $data->toArray());
-            } else {
-                $array = [];
-                foreach ($data as $name => $value) {
-                    $array[$name] = $value;
-                }
-                $this->buildXml($child, $array);
-            }
-        } elseif (is_array($data)) {
+        if (is_array($data) ||
+            ($data instanceof \Traversable && $this->useTraversableAsArray && !$data instanceof Arrayable)
+        ) {
             foreach ($data as $name => $value) {
                 if (is_int($name) && is_object($value)) {
                     $this->buildXml($element, $value);
@@ -104,6 +105,18 @@ class XmlFormatter extends Object implements FormatterInterface
                     $element->appendChild($child);
                     $child->appendChild(new DOMText((string) $value));
                 }
+            }
+        } elseif (is_object($data)) {
+            $child = new DOMElement(StringHelper::basename(get_class($data)));
+            $element->appendChild($child);
+            if ($data instanceof Arrayable) {
+                $this->buildXml($child, $data->toArray());
+            } else {
+                $array = [];
+                foreach ($data as $name => $value) {
+                    $array[$name] = $value;
+                }
+                $this->buildXml($child, $array);
             }
         } else {
             $element->appendChild(new DOMText((string) $data));