+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Yii Framework 2 HTTP client extension Change Log
 2.0.1 under development
 -------------------------
 
-- no changes in this release.
+- Bug #44: Fixed exception name collision at `Response` and `Transport` (cebe)
+- Bug #45: Fixed `XmlFormatter` unable to handle array with numeric keys (klimov-paul)
+- Enh: Added `XmlFormatter::useTraversableAsArray` allowing processing `\Traversable` as array (klimov-paul)
 
 
 2.0.0.1 July 01, 2016